pyproject.toml
[project]
name = "fun_plus_plus"
description = "The goldy bot extension with 🎉 fun inside."
authors = [
    {name = "Goldy", email = "goldy@devgoldy.xyz"},
]
readme = {file = "README.md", content-type = "text/markdown"}
requires-python = ">=3.8"
license = { file = "LICENSE" }
keywords = [
    "fun++", "fun plus plus"
]
dependencies = [
    "GoldyBot>=5.0dev12",

    "Pillow",
    "devgoldyutils>=2.5.6"
]
version = "1.0.0"

[project.optional-dependencies]
dev = [
    "ruff"
]

[project.urls]
GitHub = "https://github.com/Goldy-Bot/fun_plus_plus"
BugTracker = "https://github.com/Goldy-Bot/fun_plus_plus/issues"